fix: Ensure existence of required file provider extension database directory.
This refactors and consolidates the initialization of the database manager. It includes verification and creation of the required directory Sometimes the file provider extension crashed because it could not create the Realm database files because of the assumed directory not existing. The directory is directly related to and can be considered owned by the database manager. Hence it should also manage it.

The logging has been extended and assertion failures introduced to pinpoint irrecoverable dysfunction.

The tests have been updated accordingly and now also use unique temporary working directories by default.

Signed-off-by: Iva Horn <iva.horn@icloud.com>